Read FEI1 extended headers from MRC files.
Combine FormatMRCimage and FormatMRCstack into a single module and
derive from the same base class FormatMRC. This makes it easier to
install the MRC readers using dxtbx.install_format.

Use code from Lingbo Yu to read the MRC extended header, if present.
Currently this information is not used to set the geometry.